PSDLE - PS TV region locking.
To avoid confusion for unsupported web store languages (anything but
"en-us"), hide the option.

This is a limitation of the web store in other languages, not PSDLE.

It sucks, I know! A potential solution would be an online database of
IDs and TV compatibility. Somebody get on that.

Can now translate PS TV stuff! (...)