PSDLE - Prevent Catalog edge-cases.
Certain content could be return with a platform value of ["PS3", "PS4"
by Catalog which would break them in the system filters.

This may not be entirely accurate, but it's sane. Moving forward,
instead of breaking I expect the items to simply have the wrong system
"which is better than nothing".