PSDLE - Remove switches.
Remove switches support, moved notes to appropriate wiki page.
Ultimately it was used for one actual new switch, and a proxy to an
otherwise disabled one under certain circumstances.

Also start of climb to clean up/rewrite the startup screen which is
probably the oldest and untouched legacy code that has been ducktaped to
keep up. I intend on keeping the look of it though.