PSDLE - Clean up some translation keys.
There was a mix of spaces as underscores and camel-case. Now it's mostly
camel-case (export_view -> exportView). Also saves a few bytes.

However, we can't touch the categories keys as those are returned by the
Catalog API. If we wanted to we'd have to normalize them when we read
them, but that's a minor inconvenience upon itself.