Some more Chrome changes.
The culprit ended up being the document scroll event.
More so vendor code really getting around and I'm not about to find out what it's doing. Seems fine when bound specifically to the container, but now I have to figure out accessible scrolling or an alternative way to trigger smartScroll.