PSDLE - Two new export properties.
indexPSDLE for what its index should be based on from the valid content check, and indexRaw for the raw Entitlements index. The main difference is Raw may have gaps due to omitted services, videos, expired, or misc entitlements.

The original index property was used for internal sorting and not reliable despite the name of the new indexPSDLE property.