PSDLE - Fix includeExpired
The results are still less than ideal since older entries omit entitlement data and get relegated to generic instead. Specifically non-PS4 content.

The information we would need to forge (even if "N/A") is size and platform, and maybe copy the item/product ID into the other. "N/A" size is already done, and while they can safely all have a platform of PS3 it'd be ingenuine.

Also add the key to the base config for tab completion purposes!